Fix writing of an attribute.

+++ classfile/src/main/kotlin/com/github/netomi/bat/classfile/attribute/Attribute.kt
@@ -48,7 +48,7 @@ abstract class Attribute protected constructor(open val attributeNameIndex: Int)
 
     @Throws(IOException::class)
     override fun write(output: ClassDataOutput) {
-        output.writeByte(attributeNameIndex)
+        output.writeShort(attributeNameIndex)
         writeAttributeData(output)
     }
